added the foldesr to manage all the programs
