+++ Beginner-level- java programs/AsciiValue.java
@@ -1,3 +1,6 @@
+// these all the basic java programs come under beginner level
+
+
 // Find ASCII value of a character
 
 public class AsciiValue {
