+++ Beginner-level- java programs/BinaryToDecimal.java
@@ -1,3 +1,7 @@
+// these all the basic java programs come under beginner level
+
+
+
 class BinaryToDecimal {
     
   public static void main(String[] args) {
