+++ Beginner-level- java programs/OddEven.java
@@ -1,3 +1,5 @@
+// these all the basic java programs come under beginner level
+
 public class OddEven {
         public static void main(String[] args) {
          int n = 68;
