+++ Beginner-level- java programs/java program.java
@@ -1,0 +1,13 @@
+
+// these all the basic java programs come under beginner level
+
+class FactorialExample{  
+ public static void main(String args[]){  
+  int i,fact=1;  
+  int number=5;//It is the number to calculate factorial    
+  for(i=1;i<=number;i++){    
+      fact=fact*i;    
+  }    
+  System.out.println("Factorial of "+number+" is: "+fact);    
+ }  
+}
