+++ Beginner-level- java programs/javaleap.java
@@ -1,3 +1,7 @@
+// these all the basic java programs come under beginner level
+
+
+
 public class Main {
 
   public static void main(String[] args) {
