+++ Beginner-level- java programs/palindrome/palindrome-2.java
