+++ Beginner-level- java programs/twoSum.java
@@ -1,4 +1,6 @@
 
+// these all the basic java programs come under beginner level
+
 class Solution {
     public int[] twoSum(int[] nums, int target) {
      int i,j;
