+++ intermediate_level java program/MergeArray.java
