+++ intermediate_level java program/SettingKthBit.java
