+++ intermediate_level java program/dublicateInArray.java
